Add capability to get old meals from canteens
When requesting all meals, canteens/:id/meals is requested. That
only returns meals from the day it was requested on and the future.
However, old data can be interesting as well. With this commit, the
meals from the previous week (seven days) are requested using
the /canteens/:id/days/:day/meals endpoint. Additionally, the
Canteen.getMeals function will now request meals for the specified
date, if it is not included in the map returned by the default request.